+++ src/main/java/com/teammoeg/frostedheart/content/climate/player/TemperatureUpdate.java
@@ -173,7 +173,7 @@ public class TemperatureUpdate {
                 //Day-night temperature
                 float skyLight = world.getChunkProvider().getLightManager().getLightEngine(LightType.SKY).getLightFor(pos);
                 float gameTime = world.getDayTime() % 24000L;
-                gameTime = gameTime / (200 / 3);
+                gameTime = gameTime / ((float) 200 / 3);
                 gameTime = MathHelper.sin((float) Math.toRadians(gameTime));
                 envtemp += bt;
                 envtemp += skyLight > 5.0F ?
@@ -233,7 +233,7 @@ public class TemperatureUpdate {
                     player.attackEntityFrom(FHDamageSources.HYPERTHERMIA_INSTANT, (dheat) * 10);
                 else if (dheat < -0.1)
                     player.attackEntityFrom(FHDamageSources.HYPOTHERMIA_INSTANT, (-dheat) * 10);
-                current += dheat * tspeed;
+                current += (float) (dheat * tspeed);
                 if (current < -10)
                     current = -10;
                 else if (current > 10)
