+++ src/main/java/com/teammoeg/frostedheart/content/temperature/SteamBottleItem.java
@@ -72,7 +72,7 @@ public class SteamBottleItem extends FHBaseItem implements IHeatingEquipment, IT
 
     @Override
     public float getHeat(ItemStack is, float env) {
-        return this.getEnergyStored(is) / 120;
+        return (float) this.getEnergyStored(is) / 120;
     }
